refactor: modified flask routes
