+++ src/pyflask/collectionsDataset/__init__.py
@@ -1,7 +1,4 @@
 from .collectionsDataset import (
     get_all_collections,
-    get_current_collection_names,
-    upload_collection_names,
-    remove_collection_names,
     upload_new_names
 )
