+++ src/pyflask/datasets/__init__.py
@@ -1,1 +1,1 @@
-from .datasets import get_role, get_dataset_by_id
+from .datasets import get_role, get_dataset_by_id, get_current_collection_names, upload_collection_names, remove_collection_names
